[UsdGeom] Add an optional transform matrix parameter to the UsdBoundable::ComputeExtentFromPlugins method. If the transform is given, the extent returned is the boundable object's extent after the transform is applied to it.
Note that this transform is an arbitrary 4x4 matrix. We assume these transforms are affine, but this means they can still include non-uniform scaling and skews. This makes tight computations difficult, even for simple objects like spheres and cylinders, so we simplify by finding the extent of their transformed bounding boxes instead. For points-based objects, we find the extent of the transformed points first, then increase it by the extents of the object at each point transformed by the matrix without the translation. This gives us tighter bounds on these objects while ensuring the extents are not too small.

(Internal change: 1830345)
(Internal change: 1830765)