[usdMaya/pxrUsdMayaGL] explicitly mark the proxy shape geometry dirty when its plugs are dirtied
In this change, we explicitly notify the Maya renderer that the proxy
shape's geometry is dirty and needs to be redrawn when any plug on the
proxy shape is dirtied. We also explicitly specify isAlwaysDirty=true on
the draw override for the proxy shape.

Since isAlwaysDirty=true is the default for MPxDrawOverrides, this change
has no real effect yet, but it sets us up for when the batch renderer
eventually keeps a persistent set of shapes to be rendered from frame to
frame and we finally set isAlwaysDirty=false on the draw override.

(Internal change: 1830748)
(Internal change: 1830852)