[pxrUsdMayaGL] add shape adapter method for updating visibility
When a Maya shape is made invisible, it may no longer be included in the
"prepare" phase of a viewport render (i.e. we get no getDrawRequests() or
prepareForDraw() callbacks for that shape). The new UpdateVisibility()
method can be called on demand to ensure that the shape adapter is updated
with the current visibility state of the shape.

This will become necessary for correctly respecting Maya visibility state
when the batch renderer maintains a set of shape adapters frame-to-frame
and does not rely on the draw override/shapeUI re-queueing it at every
draw.

(Internal change: 1830869)