Add changed fields to UsdNotice::ObjectsChanged
Users can now query the changed fields that affected
the objects reported in the UsdNotice::ObjectsChanged
notice. This allows for more selective change processing
in client code.

The GetResyncedPaths and GetChangedInfoOnlyPaths functions
now return a range object instead of an SdfPathVector.
The iterators for this object can be used to access the
changed fields for each path. Convenience API for this
information was also added to the ObjectsChanged notice
itself.

To minimize the cost for providing this extra information,
the change processing code in UsdStage was reworked to
avoid unnecessary copies and unpacking of sets to vectors.

(Internal change: 1831411)