[pxrUsdMayaGL] make the batch renderer persist its bucketed shape adapters frame-to-frame
This change moves the batch renderer away from maintaining render and
select queues of shape adapters that are swapped/dumped at each
render/selection cycle and instead maintains a persistent set of bucketed
shape adapters across frames.

Previously, we used the presence of elements in the render or select queues
to determine when to perform a batched operation and when to pass through
or re-use cached data. We now use a different mechanism for this:
    - For Viewport 2.0, we use the draw context's frame stamp to identify
      the draw/selection operation being performed. This ensures that we
      only execute one batched draw or batched selection for a given
      operation.
    - For the legacy viewport, frame stamps are not available, so we
      simulate them by setting a render pending flag when
      CreateBatchDrawData() is called (indicating a draw operation is
      coming), and setting a selection pending flag when a render
      completes.

A new Tf debug code, PXRUSDMAYAGL_SHAPE_ADAPTER_BUCKETING, was added for
debugging the bucketing process as shape adapters are added to and removed
from the batch renderer. PXRUSDMAYAGL_QUEUE_INFO will be renamed and/or
broken up in a subsequent change.

This change should also allow us to start taking advantage of
isAlwaysDirty=false in MPxDrawOverrides.

(Internal change: 1831613)