Remove pre-authoring workaround in usdview
Now that usdImaging no longer resyncs in response
to adding empty overs on prims, the workaround in
usdview that pre-authored overs in a session layer
can be removed. This means we no longer have to
open the stage twice (the first open was generally
cheaper since no models were loaded, but was
still not free).

(Internal change: 1831781)