[pxrUsdMayaGL] clean up MUserData draw data handling and allow for reuse in Viewport 2.0
In Viewport 2.0, MPxDrawOverride::prepareForDraw() may receive a pointer to
an existing MUserData object that can be reused for a subsequent draw.
Before this change, we were not taking advantage of that and were instead
creating a new object at every call.

There is no change in pattern for the legacy viewport. The draw data is
always created in CreateBatchDrawData(), and is destroyed at the end of
the legacy viewport Draw().

(Internal change: 1831788)