Copied over PcpComptePrimPropertyNames() and its dependencies as a private helper function in usd/prim.cpp and optimized it for USD. The new function no longer maintains a PcpTokenSet which has O(n^2) insertion time. It simply keeps adding to the vector and we do the uniquification and ordering once at the end.
In addition to the optimization, the new function also takes a predicate function for filtering
properties. Only the properties whose names pass the predicate are included in the result.

UsdPrim::Get{Authored}Properties and UsdPrim::Get{Authored}PropertyNames() now take an
optional 'predicate' function parameter. When a valid predicate is passed in, they now collect and
return only the properties that pass the predicate.

Updated implementation of UsdPrim::GetPropertiesInNamespace() to take advantage of the
"predicate" parameter and only collect property names that are in the requested namespace.

(Internal change: 1831820)