[pxrUsdMayaGL] set isAlwaysDirty=false on the USD proxy shape MPxDrawOverride
With the set of shape adapters registered with the batch renderer now being
persisted from frame to frame, we can set isAlwaysDirty=false when
constructing the UsdMayaProxyDrawOverride. This will have the effect of
prepareForDraw() only being called when something on the shape actually
changes or when the display state changes, rather than at every display
refresh.

(Internal change: 1831952)