[HdSt] Add texture state to surface shader hash.
There is an internal test that disconnects and reconnects texture bindings.  This test was failing removing the binding and was displaying black instead of the expected color.  The test uses a custom surfaceShader, that is if the texture is unbound, it will fallback to the color passed in to the  surfaceShader terminal.

The surface shader hash is used to determine if the code generator needs to be re-run when building the draw batches.  However, the texture state wasn't being taken into account, so the shader was reading from the unbound texture and getting the api's default texture color, instead of taking the fullback path.

The fix for this issue to add texture state to the surface shader hash, so that the code gets re-generated and the fullback path is used.

Changes:
- In HdStSurfaceShader::ComputeHash(), include the number of textures, the texture name and texture type in the shader hash.

(Internal change: 1832028)