[usdMaya] Redo the skel test Maya scene in Maya 2016 to ensure compatibility
skinCluster weights are authored in 2018 in a way that's not backwards-compatible with 2016.
I opened the file in 2016, unbound the skinClusters, rebound them at the joint rest pose,
then moved the joint backs to their posed positions. I then re-saved the file in 2016 and manually
removed cruft nodes. It now opens in 2016 without error.

(Internal change: 1832290)