[pxrUsdMayaGL] fix for capturing transform changes in Viewport 2.0
With the move to use isAlwaysDirty=false in the MPxDrawOverride for USD
proxy shapes, we now only get a prepareForDraw() call when the proxy shape
itself is dirtied. This does not include any changes in the shape's
transform, which actually comes from the shape's parent node rather than
the shape itself. To capture those changes, we intercept requests for the
transform in the draw override and propagate those to the shape adapter's
delegate.

(Internal change: 1832295)