[pxrUsdMayaGL] more separation between legacy viewport and Viewport 2.0 shape adapters/delegates
With this change, the batch renderer now maintains two separate SdfPath
hierarchies of shape adapter/delegate IDs, one branch for the legacy
viewport and one branch for Viewport 2.0. The MAYA_VP2_USE_VP1_SELECTION
environment variable still throws a bit of a wrench in the works, but we're
better able to deal with it now by simply re-writing the delegate ID we're
testing for intersection with when we're in a legacy-based selection for
Viewport 2.0. This resolves an issue where switching back and forth between
viewport renderers would cause one renderer or the other to seemingly not
redraw correctly in response to shape/transform changes. This was the
result of two separate UsdImagingDelegates with the same ID trying to
manage the same set of rprims. We now still have two separate delegates
(one owned by the MPxSurfaceShapeUI/legacy viewport shape adapter, and one
owned by the MPxDrawOverride/Viewport 2.0 shape adapter), but they have
unique delegate IDs, and therefore unique rprims.

This will also come into play when we re-collapse the rprim collections for
selection, and eventually also for drawing. All shape adapters/delegates
for a particular viewport can be addressed by a single collection with the
viewport-appropriate prefix as the root path.

A new "common" _Sync() method was added to the shape adapter class that is
intended to be called by the public, viewport-dependent Sync() methods.
This allows us to have an accurate value for _isViewport2 at the time that
we're initializing the shape adapter, and is also just generally easier to
follow than one public Sync() calling into the other. The new _Sync()
method is a virtual protected method looking forward to when we have
multiple subclasses of shape adapter for different Hydra-image Maya shape
types.

This change also adds a public token for the Maya type name of USD proxy
shape nodes. We use this to build the delegate name in the shape adapter to
help avoid name collisions between Hydra-imaged shapes of different Maya
types.

(Internal change: 1832426)