Optimizations to UsdShadeMaterialBindingAPI::ComputeBoundMaterial:
* When computing direct and collection-based bindings on a prim, call GetPropertyNames once
with a predicate that only passes for properties whose names start with "material:binding". If this
results in an empty vector there can be no bindings to process on the prim. We can also use this
to determine whether direct or collection-based bindings exist for the current value of material
purpose.

* Converted CollectionQueryCache to a tbb::concurrent_unordered_map<SdfPath, unique_ptr<CollectionQuery>>
This makes parallel population and traversals of the cache threadsafe.

* Added another cache for memoizing per-prim direct and collection-based bindings. This is also
implemented using a tbb::concurrent_unordered_map, which makes it threadsafe as long as entries
aren't erased from the map concurrently. This cache gets populated lazily as more and more bindings
are resolved on prims, thereby avoiding redundant binding computations on ancestor prims during
initial population, and on leaf prims (on subsequent binding resolutions).

* Updated UsdImaging code to take advantage of the two caches in its computation of resolved
material bindings.

(Internal change: 1832811)
(Internal change: 1832841)