Add an optional 'instanceName' argument to UsdPrim::HasAPI(), which can be used to query if a prim has had a particular instance of a multiple-apply API schema applied to it. If "instanceName" is empty and the API schema type is a multiple-apply API schema, UsdPrim::HasAPI now returns true if the prim has one or more instances of the API schema applied to it.
The "instanceName" argument value should be empty if the API schema type is
a single-apply API schema. If not, a coding error is issued.

To enable the above behavior:
* added the names of all the multiple-apply API schemas to the layer customData in
the generatedSchema.usda files. The schema registry now populates these names in
a hash-set and provides API to query whether a given API schema (identified its
TfType) represents a multiple-apply API schema.
* as part of codegen, added a compile-time constant "IsMultipleApply" to all API schemas
that indicates whether the API schema is multiple-apply or single-apply.

(Internal change: 1832812)