[HdSt] Updated HdStExtCompuGpuComputation
The main functional change here is to maintain a clearer
distinction between data buffers belonging to the computation
and data buffers belonging to the buffer array range of a
host rprim.

This has two main implications:
- first this fixes a bug that prevented the output of
  a computation from having a different name than
  a computed primvar using the output
- second this allows a computed primvar to be
  processed consistently by other rprim computations,
  e.g. smooth normals, and subdivision refinement.

This also reorganizes a bit of the setup code to eventually
improve the interaction between CPU and GPU computations.

(Internal change: 1833077)