[usdMaya] sync the usdWriteJob export loop between file exports and the usdExport command
Before this change, there was a subtle difference in behavior between doing
a "File -> Export" out to USD versus running a usdExport command. With the
former, it did not consider the "exportAnimation" job arg and would always
invoke usdWriteJob::evalJob() at least once, even if it was only for a
single frame. With usdExport however, usdWriteJob::evalJob() would *not*
get called if "exportAnimation" was false.

This change syncs the "File -> Export" behavior with usdExport so that in
both cases, usdWriteJob::evalJob() is only called for ranged exports.

(Internal change: 1835742)