[pxrUsdMayaGL] optimize single selection by maintaining viewport renderer-based collections
In the case of single selections (as opposed to area/marquee selections),
we're only interested in the single nearest hit object. We can avoid a lot
of unnecessary work by testing for intersection against all of the geometry
represented by all of the shape adapters registered with the batch renderer
at once, rather than testing each shape adapter's collection one at a time.
We continue to do the latter for area selections though, as we want to
support the functionality where occluded objects are still included in the
resulting selection.

With this change, the batch renderer holds an rprim collection for each
viewport renderer. When computing selections, the appropriate collection is
used for single selections.

This change also migrates management of the culling style to the individual
shape adapter delegates rather than being a render param that is translated
into a parameter on the HdxRenderTaskParams.

(Internal change: 1835787)