[usdMaya] add a dedicated prim writer for Maya locators
This change adds a simple prim writer specifically for Maya locator shape
nodes. This ensures that locators are exported correctly whether
mergeTransformAndShape is turned on or off, and it avoids further
complicating the logic for node collapsing and exporting transforms in
MayaTransformWriter.

Note that there is currently no "Locator" type in USD and that Maya locator
nodes are exported as UsdGeomXform prims. This means that locators will not
currently round-trip out of Maya to USD and back because the importer is
not able to differentiate between Xform prims that were the result of
exporting Maya "transform" type nodes and those that were the result of
exporting Maya "locator" type nodes.

This change also adds a safety check in usdWriteJob::evalJob() to ensure
that each prim writer has a valid prim before calling the prim writer's
write() method. This was already happening for the default time export in
usdWriteJob::beginJob(). This avoids a crash that could happen for frame
ranged exports where a prim writer is valid but for whatever reason does
not define a prim during the default time export (as was the case before
adding the locator prim writer).

Fixes #417, Fixes #129

(Internal change: 1836043)