One more step in removing a direct dependency on hdSt from usdImagingGL.
This change adds the ability for each renderer plugin to perform a
runtime check to make sure that the minimum requirements are met before
the renderer is invoked.  This is done through a new virtual IsSuppored
on HdxRendererPlugin.

Moved the concept of "default renderer plugin" to the Hdx registry from
the Hf registry, because it can now check to find the first *supported*
plugin as the default (not just the first one).  Here we are very careful
to not load more plugins than necessary.

Up until this change, if a platform didn't support hdStream level of OpenGL
then none of the backends were available to it, not even ones that are
purely CPU based and don't require an advanced GL at all.  With this change
we will now search through the backends and find the first one that is
supported.  Only if none are supported do we say that "Hydra" is not
supported and fallback to the reference renderer in usdview.

(Internal change: 1836060)