include documentation for code in third_party in doxygen build
This also includes a re-ordering of libraries and plugins listed as INPUTs
in the master Doxyfile to align the ordering more closely with the desired
lowest-to-highest level order.

A few minor issues in the overview.dox files in third_party were fixed to
get the top-level pages to show, and to fix some incorrect symbol names and
patch up some broken links.

(Internal change: 1836194)