Sparse invalidation of usdImaging for prim changes
UsdImagingDelegate now defers to the various prim adapters
to determine if a prim resync is necessary in response to
non-structural scene description changes.

By default, a resync is considered necessary if a change
was made to any plugin metadata fields; changes to built-in
metadata fields are ignored. This behavior was selected
because schemas typically register plugin metadata fields
to control their behavior rather than rely on the built-in
fields.

This change makes it so that changes to fields that aren't
relevant to imaging (e.g. 'documentation') won't cause
unnecessary resyncs in usdImaging.

(Internal change: 1836365)
(Internal change: 1836426)