[UsdGeom] Added multisampling support for UsdGeomPointInstancer with a new ComputeInstanceTransformsAtTimes method.
ComputeInstanceTransformsAtTimes does the same operation as ComputeInstanceTransformsAtTime, but takes multiple sample times and returns a vector containing the resulting instance transforms for each. This new method is more efficient than multiple calls to the single-sample method because data is only fetched once across multiple samples. Because much of the code is the same for these two methods, each calls a new private preamble method which fetches the data they both need.

The katana plugin had already implemented this functionality and it was updated to use this new, more standard method. The plugin originally took a velocity scale in its opArgs which is not handled by ComputeInstanceTransformsAtTimes, so sample times are pre-scaled using this custom velocity scale and a deprecation warning is given if the velocity scale is specified. Future velocity scales should use an authored 'motion:velocityScale' attribute in USD.

(Internal change: 1836995)