[Usdview] Fixed broken menu items.
Menu items modified in change 1837055 were broken because their callback methods were still expecting a 'checked' boolean to be passed as a parameter (this is from the 'toggled' signal, and these items were updated to use the 'triggered' signal which passes no parameters).

Some menu items modified in change 1836602 became locked and could not be changed because their 'triggered' callback methods set multiple view settings. After each setting was changed, the UI was refreshed to match the view settings state which can overwrite the intended change in the UI before it can be used to actually update the setting.

(Internal change: 1837784)