Bump minimum TBB version to 4.4 Update 6
This brings the USD requirements in line with the
VFX reference platform for CY2017 and allows
code to make use of API added to the concurrent
containers starting with 4.3 Update 4.

Documentation, CI build configuration, and
build_usd.py build script have all been updated
to reflect this change.

(Internal change: 1838125)