[UsdMaya/PxrUsdMayaGL] Maya live surface support for USD proxy shapes
In UsdMaya, the UsdMayaProxyShape will hold a delegate that is used to compute
the live-surface intersection point. (Currently, we don't compute the live-surface
normal, always specifying the Maya system up-vector instead.)

In PxrUsdMayaGL, a delegate is defined for the UsdMayaProxyShape, and registered
via a registry function. The delegate uses Hydra rendering of a camera pointed along
the live-surface ray in order to provide the intersection.

The batch renderer's draw and intersection routines are slightly refactored to enable
drawing a custom collection with a single shape, as opposed to drawing all shapes in
a batch. Namely, the modifications are:
    - Provide an API to populate a custom collection from the UsdMayaGLBatchRenderer,
      and Draw/TestIntersection overloads that deal with custom collections.
    - Provide an extra parameter on PxrMayaHdRenderParams to enable custom bucketing.
      This is because we don't want to reuse the same render tasks used for batch rendering
      and custom single-shape rendering; that would cause the render tasks to repeatedly
      dirty since their rprim collections will keep changing. Instead, we want to make sure
      that any custom rendering is done using a separate "bucket" for each custom
      collection.
    - Expose the DAG path of shape adapters, and keep a map of DAG path to shape
      adapters for both VP2 and legacy viewports. Currently, the custom collection
      Draw/TestIntersection methods are agnostic to whether their shape adapters come
      from VP2 or Legacy -- they'll prefer VP2 first but check Legacy if VP2 adapters aren't
      available. This is because they are designed for drawing into a custom target instead
      of the Maya viewport.

(Internal change: 1838840)