[pxrUsdMayaGL] fix for batch renderer-related crashes when doing "File -> Save Scene As..."
The batch renderer needs to be reset when changing scenes (either by
switching to a new empty scene or by opening a different scene). We listen
for kAfterNew and kAfterOpen scene messages specifically and *not* for
kSceneUpdate messages since kSceneUpdate is also emitted after SaveAs
operations, in which case we actually do not want to reset the batch
renderer.

(Internal change: 1839020)