[UsdMaya] Make instances work better with model hierarchy
The usdWriteJobCtx now stores a mapping between Maya instanced objects and their prim writers.
The MayaTransformWriter will now forward the exportsGprims() and exportsReferences() functions
to the instance master prim writer's implementations (where previously they returned false).
This causes the ModelKindWriter to properly validate model hierarchy on prims that are instanced.

In addition, the processInstances() invocation is now moved after all prims should have been defined,
but before the ModelKindWriter runs. Since processInstances() forces the InstanceSources prim to
be an over, it needs to run before ModelKindWriter so that the ModelKindWriter knows to skip it
during model hierarchy authoring and validation.

(Internal change: 1839659)