Added API to support the following standardized material terminal outputs: surface, displacement and volume.
* Added "outputs:ri:surface", "outputs:ri:displacement" and "outputs:ri:volume" as built-in
token-valued attributes on a Material prim. These represent the outputs corresponding to the
universal render-context (which is represented by the empty valued token
UsdShadeTokens->universalRenderContext).
* Added Get<Name>Output() and Create<Name>Output() methods on UsdShadeMaterial
to get and create terminal outputs for a given render-context.
* In addition to the Get and Create methods, there's a Compute<Name>Source()
method for each of the three outputs. This method computes the connected source
shader for a specific render-context. If the output corresponding to the requested renderContext
does not exist or is not connected, the source of the universal renderContext is returned
by this method.

As part of this change,
* Made the variability of a UsdShadeOutput 'uniform', since its value should
never be animated. In fact, it is unusual to even set a value on an output.
* Better const correctness for a lot of methods in UsdShade.
* Added UsdShadeNodeGraph::ComputeOutputSource(), to recursively
resolve the source of a node-graph output to a shader output.
* Removed the deprecated ri:bxdf output and all relevant API from UsdRiMaterialAPI.
UsdRiMaterialAPI::GetSurface() handles the backwards compatibility by looking for
the "ri:bxdf" source if it doesn't find a "ri:surface" source.
* The various Get<output>() methods on UsdRiMaterialAPI now return a
generic UsdShadeShader object, as we're moving away from specific concrete
shader prim types.
* Updated various clients to read and author the "ri:surface" output instead of the
"ri:bxdf" output.

(Internal change: 1840047)