[Hd,HdSt] ExtComputation Sprim
Changed HdExtComputation to be an HdSprim subclass instead
of implementing special handling.

This allows render delegates to specialize behavior following
existing patterns, e.g. to allocate and manage GPU buffer storage.
One limitation of this change is that it doesn't specifically support
management of imaging computations from populated from one
scene delegate to serve as inputs to an rprim populated by a
different scene delegate. This feels like something that we will
want to revisit eventually, but does not affect current use cases.

Additionally, this introduces a special semantic for aggregation
of imaging computation inputs. When an imaging computation
is declared with no outputs and no computation kernels, it acts
as an input aggregation and allows a render delegate to commit
GPU resources for a set of inputs separately from other inputs.

(Internal change: 1840224)
(Internal change: 1840302)