[UsdGeom] Added multisampled extent computations for point instancers with ComputeExtentAtTimes method.
The existing ComputeExtentAtTime method was refactored so it could share code with this new method. Both use a preamble method which fetches common data and a function which actually computes extents which is called once for each timesample. Also, in several places 'time' was used where 'baseTime' should have been used (this is done to guarantee the point instancer does not change topology between samples).

testUsdGeomComputeInstanceTransformsAtTime was renamed to testUsdGeomComputeAtTime and was updated with new tests for both the single sample and multisample versions of ComputeExtentAtTimes.

(Internal change: 1840370)