Provide explicit C++ typenames for certain scene description types.
Scene description value types may now be associated
with a C++ typename that overrides the typename
provided by TfType. The latter is based on the
underlying C++ type, which is problematic if the
user-facing type is really a typedef, like int64_t.
As part of this change, value types that run afoul
of this issue have had explicit typenames added.

These typenames are primarily used for
documentation and diagnostics. An upcoming change
will update USD schema generation to take advantage
of this information.

(Internal change: 1840505)