[usdMaya] Fix crash in usdReadJob::DoImport caused by dangling references/pointers in prim readers.
Prim readers hold onto a reference to their args and a pointer to the context.
But when prim readers are saved, the args and context are not, so the
pointers/references become invalid. This change makes prim readers (1) store
a copy of their args instead of a reference, and (2) not store any pointer to the
context. Since the context is transient and tied to specific invocations of
Read/PostReadSubtree, a fresh context is provided by the usdReadJob.

(Internal change: 1841167)