Adding trace library and removing tracelite
The trace library provides a standard, built-in
platform for capturing and reporting performance
in a variety of formats. This library has been
used at Pixar to profile code for many years.
By including it with USD, we hope to provide a
useful tool for measuring USD performance as
well as a common way to share this data for
diagnostic purposes. Much of the USD codebase is
already instrumented in key locations, so users
can immediately begin measuring performance in
their use cases.

The tracelite library provided stubs that allowed
clients to hook in their own profiling code. With
the release of the trace library, this library is
no longer needed.

(Internal change: 1843848)
(Internal change: 1844400)