[pxrUsdMayaGL] fix for incorrect ambient being applied in the legacy viewport
This change fixes an issue where shapes imaged by the Maya/Hydra batch
renderer may incorrectly have a (0.2, 0.2, 0.2, 1.0) value applied to their
lighting when drawn in the legacy viewport.

Maya does not appear to use GL_LIGHT_MODEL_AMBIENT, but it leaves the
default value of (0.2, 0.2, 0.2, 1.0) in place. The first time that the
viewport is set to use lights in the scene (instead of the default lights
or the no/flat lighting modes), the value is reset to (0.0, 0.0, 0.0, 1.0),
and it does not get reverted if/when the lighting mode is changed back.

Since in the legacy viewport we get the lighting context from OpenGL, we
read in GL_LIGHT_MODEL_AMBIENT as the scene ambient. We therefore need to
explicitly set GL_LIGHT_MODEL_AMBIENT to the zero/no ambient value before
we do, otherwise we would end up using the "incorrect" (i.e. not what Maya
itself uses) default value.

This is not a problem in Viewport 2.0, since we do not consult OpenGL at
all for any of the lighting context state.

As an extra precaution, the scene delegate also now detects whether the
lighting context's scene ambient or material have changed versus what was
last set in the lighting task's parameters and correctly sets the
parameters dirty if so.

(Internal change: 1843772)