[usdMaya] fix for USD proxy shape attributes with connections not causing redraws
A number of input attributes on the USD proxy shape do not have an impact
on the USD stage data output attributes of the shape, but they *do* have
an affect on how the shape is imaged in the viewport. When these attributes
get their value from an upstream connection, we may not get a
setDependentsDirty() callback for the plug and may instead only get a
compute() callback. Before this change, that would mean that when using
Viewport 2.0, we were not always notifying Maya that the shape needed to be
redrawn, and so the viewport might not reflect changes to those attributes.

With this change, if we receive a compute() callback for one of these
attributes, we notify Maya that the shape needs to be redrawn and then let
Maya take care of computing the attribute.

(Internal change: 1843808)