[HdSt] Fixed over-invalidation issue on material parameter change.
There's an dependency in Hydra, where an Rprim's topology depends on if the Rprim's material has a ptex bound to it.
To solve this problem in a previous checkin, the material would dirty all Rprim's when the material params changed, so that they can re-evaluate their state.

However, this proved to be too big a hammer, as dirtying the material binding also causes all the primvars to be refetched and that was far too expensive for every rprim in a large scene.

As ptex binding changes are really rare in practice, this change limits the use of the hammer to just a ptex state change.

Changes:
- Added a flag to material, to track the current state of if a ptex is attached to the material.
- Only set the Material id dirty bit on rprims if the ptex state is different.
- HdStMesh now takes advantage of the summary bit on the meterial, rather than having to walk all params on the shader to look for it.

(Internal change: 1843832)