[pxrUsdMayaGL] fix handling of shape visibility with display layers
The UpdateVisibility() method on the USD proxy shape adapter was previously
only using MHWRender::MGeometryUtilities::displayStatus() to determine
whether its shape was visible or not. This does not take into consideration
the visibility of any display layers that the shape (or one of its
ancestors) may be a member of. With this change, we also query the shape's
dag path for its visibility state when updating visibility.

(Internal change: 1843865)