[pxrUsdMayaGL] gate area depth selections behind an env setting
When doing an area/marquee selection with native Maya geometry, Maya
selects objects that are occluded by other objects in depth if the
selection would have otherwise included the occluded object. We want to
provide the same behavior in the selection implementation of shapes imaged
by Hydra, but that currently comes with a significant performance penalty
if the number of objects grows large.

For now we only expose that behavior when the following env setting is
enabled:

    PXRMAYAHD_ENABLE_DEPTH_SELECTION

(Internal change: 1843993)