[pxOsd] verify crease data in refinerFactory during tag assignment.
The data could be bad and cause segment violations due to over and underruns.
Warn about the presence of, and avoid bad creases.

FindEdge is not bounds checking, and crease data could be referencing outside
the bounds.

The asset itself may need fixing if any of the warnings fire off.

(Internal change: 1844274)