[katana] Use SdfAssetPath::GetResolvedPath instead of running resolver
This fixes relative asset names from usd files not in the current directory, and
removes redundant execution of searches and resolve.

If for any reason GetResolvedPath has not been set, the raw asset string is used
and a warning message is produced.

This patch also defaults asShaderParam to true because almost every call to
this function uses it that way, the ri statements converter is the only
exception.

Signed-off-by: Bill Spitzak <bill.spitzak@dreamworks.com>