[Hdx] Added partial support for "pick through" mode
Pick Through doesn't account for occlusion of pickable prims.

This CL extends Hydra's rasterizer-based picking solution to partially support PickThrough semantics by adding just the required configurability to HdxIntersector.
The reason for partial support is that the solution is limited by rasterization, so anytime a prim/subprim is fully hidden, we won't pick it.

Changes:
- Renamed HdxIntersector::PickMode to PickTarget, which conveys what we're attempting to pick (prim, faces, edges, points)
- Added a bool, pickThrough, to HdxIntersector::Params, which disables optional occlusion passes prior to the ID render. This allows us to mimic "pick through" workflow operations.

In the process, a latent bug was fixed:
- HdxIntersector::ResolveNearest always returned the pixel id that was nearest, w/o checking if it corresponded to a picked prim. So, we ended up missing picks for subprims such as points when the surface was treated as an occluder.

(Internal change: 1845632)