[UsdMaya] Add new "Cards" representation to USD reference assembly.
The "Cards" representation is backed by a drawMode attribute (not storable) on the assembly.
Switching to cards sets drawMode=cards, switching out of cards clears drawMode.
Whenever drawMode is non-empty, model:drawMode is set to that value, and
model:applyDrawMode=1 (in the session layer) to force the draw mode.
Since the attribute isn't storable, the drawMode isn't written to the scene file (though we are
free to change this behavior in the future).
Session layers for the nested assembly stages are now no longer pulled from a cache if the
nested assembly needs to write additional session-level opinions.
This caused problems when variant/drawMode edits were applied to a nested assembly and
the edits were applied to all assembly hierarchies using the shared session layer.
The nested assembly now automatically creates a new session layer and stage if needed.

(Internal change: 1845842)