[Hd] Fix to pickup changes to render tags
While Render Tags are currently fetched from the scene delegate every time they are needed.  Hydra does contain state that depends on the render tags.  These are Dirty List and Draw Batches, because these two items filter out prims that don't match the render tag.  Therefore, whenever the render tag changes we need to update these states.

Therefore, Render Tag needs change processing like other state.  "Purpose" was an old name for this state, but the dirty bit's name was not updated to reflect the name change.  So this change updates the naming of the dirty bit.

A key feature of this change processing, is we actually have to reprocess the dirty list state.  Therefore, the fix is in MarkRprimDirty.  When it sees the DirtyRenderTag bit, it increments the render index version, effectively treating it as a scene hierarchy edit.  This is believe to be ok as it is expected that Render Tag changes for a prim to be a very low frequency operation.

Changes:
- Renamed DirtyPurpose to DirtyRenderTag
- In MarkRprimDirty(), if DirtyRenderTag is set increment _indexVersion

(Internal change: 1846069)