[pxrUsdMayaGL] expose MUserData derived class for use with Maya/Hydra shape adapters
This change extracts the _BatchDrawUserData derived class of MUserData that
was previously internal to the batch renderer as a public PxrMayaHdUserData
class. This will allow shape adapters to create their own derived classes
of PxrMayaHdUserData if they want to store any additional data on the user
data that is not specific to the batch renderer, or if MPxSurfaceShapeUI or
MPxDrawOverride-derived classes want to be able to distinguish between
Hydra and non-Hydra user data.

(Internal change: 1847301)