[Hd] Point picking via Hydra now returns the correct point index Previously, the point id output during the ID render (picking) pass was simply the gl_VertexId. Hydra Stream uses aggressive buffer aggregation strategies, which means that it is more often the case that a non-zero baseVertexOffset is used in the glDraw* command.
gl_VertexId is the "net" vertex id, and includes this offset.
Thus, to get the actual vertex id within a draw call, we need to subtract the baseVertexOffset from it. This is available as the built-in variable gl_BaseVertex in GLSL 4.60 or gl_BaseVertexARB if the ARB_shader_draw_parameters extension is enabled. This would however mean that for cards/drivers that don't support the extension, point picking wouldn't work.
To provide an extension/version free solution, we plumb the baseVertexOffset (from vertex attrubute aggregation) as "vertexCoord" in the shader struct "hd_drawingCoord". This increases the data bound per draw call by 4 bytes. Code gen then abstracts the capabilities of the system away by using the appropriate solution within "GetBaseVertexOffset".

Changes:
1) CodeGen:
- Added vertexCoord to hd_drawingCoord, between constantCoord and elementCoord members. Internally, we generate this struct from drawingCoord0 and drawingCoord1.
  The latter has been updated to a vec4 from vec3, with vertexCoord as the .w element. This allows the vertex attribute fetch optimization in indirectDrawBatch to stay unchanged (i.e., we fetch upto instanceIndex, and skip shaderDC and (now vertexDC).
- GetBaseVertexOffset is additionally generated during vertex primvar processing.

2) ResourceBinder
- Updated drawingCoord1 to use vec4 (from vec3)

3) ImmediateDrawBatch
- Updated binding code for drawingCoord{0,1} to mirror the expectation in CodeGen.

4) IndirectDrawBatch
- Updated all *Command structs to include and plumb vertexDC at the end.
- Updated drawingCoord1 binding in the dispatch buffer.

5) pointId.glslfx
- ProcessPointId now accounts for the base vertex offset by calling GetBaseVertexOffset.

(Internal change: 1847432)